requirements.txt
langchain
langchain-community
langchain-core
plotly
pandas
psycopg2-binary
python-dotenv